+++ rouver/router.py
@@ -333,7 +333,7 @@ def _respond_not_found(environment: WSGIEnvironment,
         -> Iterable[bytes]:
     path = cast(str, environment.get("PATH_INFO", ""))
     message = "Path '{}' not found.".format(path)
-    page = http_status_page(HTTPStatus.NOT_FOUND, html_message=message)
+    page = http_status_page(HTTPStatus.NOT_FOUND, message=message)
     return respond_with_html(
         start_response, page, status=HTTPStatus.NOT_FOUND)
 
@@ -346,7 +346,7 @@ def _respond_method_not_allowed(
     message = "Method '{}' not allowed. Please try {}.".format(
         method, method_string)
     html = http_status_page(
-        HTTPStatus.METHOD_NOT_ALLOWED, html_message=message)
+        HTTPStatus.METHOD_NOT_ALLOWED, message=message)
     return respond_with_html(
         start_response, html, status=HTTPStatus.METHOD_NOT_ALLOWED,
         extra_headers=[("Allow", ", ".join(allowed_methods))])
@@ -355,7 +355,7 @@ def _respond_method_not_allowed(
 def _respond_internal_server_error(start_response: StartResponse) \
         -> Iterable[bytes]:
     html = http_status_page(HTTPStatus.INTERNAL_SERVER_ERROR,
-                            html_message="Internal server error.")
+                            message="Internal server error.")
     return respond_with_html(start_response, html,
                              status=HTTPStatus.INTERNAL_SERVER_ERROR)
 
@@ -363,7 +363,7 @@ def _respond_internal_server_error(start_response: StartResponse) \
 def _respond_http_exception(start_response: StartResponse,
                             exception: HTTPException) -> Iterable[bytes]:
     status = HTTPStatus(exception.code)
-    html = http_status_page(status, html_message=exception.description)
+    html = http_status_page(status, message=exception.description)
     return respond_with_html(start_response, html, status=status)